Property search and housekeeping
Property search function added to accept a list of properties that you wish to extract from the master list, instead of simply extracting the entire document. Ability to extract the entire document has been preserved.

Also changed file pathing to look in a folder in the parent directory, instead of directly in the subdirectory the script is in. Makes it easier to avoid pushing production information to github.

Also moved the word separation function, that actually does the extraction, to its own method for future configurations.